chore: refactor migration system
