+++ packages/background/src/index.ts
@@ -6,6 +6,12 @@ import {
 } from "./lib/context-menu";
 import { initializeStorage } from "./lib/storage";
 
+/**
+ * From the documentation:
+ *
+ * Fired when the extension is first installed, when the extension is updated
+ * to a new version, and when the browser is updated to a new version.
+ */
 browser.runtime.onInstalled.addListener(async () => {
   initializeContextMenu();
   initializeStorage();
