+++ packages/content/src/index.ts
@@ -1,8 +1,6 @@
-import {
-  renderContent,
-  runStorageMigrations,
-  startContentListeners,
-} from "./lib";
+import { runStorageMigrations } from "@worm/shared";
+
+import { renderContent, startContentListeners } from "./lib";
 
 document.addEventListener("readystatechange", () => {
   renderContent("document state change");
