+++ packages/content/src/lib/index.ts
@@ -1,3 +1,2 @@
 export * from "./listeners";
-export * from "./migrations";
 export * from "./render";
