+++ packages/shared/src/storage/api.ts
@@ -6,24 +6,13 @@ import {
   StorageVersion,
 } from "@worm/types";
 
-import { browser } from "./browser";
-import { logDebug } from "./logging";
-import { matchersFromStorage, matchersToStorage } from "./matchers";
+import { browser } from "../browser";
+import { logDebug } from "../logging";
+import { matchersFromStorage, matchersToStorage } from "../matchers";
 
 const {
   storage: { sync },
 } = browser;
-
-/**
- * The storage version from which all others are based. This should be the
- * starting point for migrations and used only when no version exists yet.
- */
-export const BASELINE_STORAGE_VERSION: StorageVersion = "0.0.0";
-
-/**
- * The storage version currently in use.
- */
-export const CURRENT_STORAGE_VERSION: StorageVersion = "1.1.0";
 
 export const storageClear = sync.clear;
 export const storageGet = sync.get;
